player class

+++ src/main/java/htl/steyr/demo/cards/PlayingCard.java
@@ -3,28 +3,28 @@ package htl.steyr.demo.cards;
 public class PlayingCard {
 
     private String name;
-    private double statistik_1;
-    private double statistik_2;
-    private double statistik_3;
-    private double statistik_4;
+    private int statistik_1;
+    private int statistik_2;
+    private int statistik_3;
+    private int statistik_4;
 
     public String getName() {
         return name;
     }
 
-    public double getStat1() {
+    public int getStat1() {
         return statistik_1;
     }
 
-    public double getStat2() {
+    public int getStat2() {
         return statistik_2;
     }
 
-    public double getStat3() {
+    public int getStat3() {
         return statistik_3;
     }
 
-    public double getStat4() {
+    public int getStat4() {
         return statistik_4;
     }
 }
